+++ django_ninja_admin/constants.py
@@ -1,0 +1,7 @@
+import enum
+
+
+class ShowFacets(enum.Enum):
+    NEVER = "NEVER"
+    ALLOW = "ALLOW"
+    ALWAYS = "ALWAYS"